refactor: Update redirect option in login page
The code changes update the redirect option in the login page to improve the user experience. After successful login, the user will now be redirected to the desired page. This change ensures that the user is directed to the appropriate location, enhancing usability.

Import `ResponseImage` in `ImageForm.tsx`

The code changes in `ImageForm.tsx` import the `ResponseImage` component from `@/(desap)/ento/calculator/page`. This import is necessary for the proper functioning of the component.

Refactor `ImageForm.tsx` to remove unused code

The code changes in `ImageForm.tsx` remove unused code related to the `LoadingComponent` and `FaFileUpload` imports. This cleanup improves code readability and maintainability.

Refactor code to use client-side rendering for council detail components

The code changes refactor the code to use client-side rendering for the council detail components. This change improves performance and provides a smoother user experience.

Update landing page link in Navbar component

The code changes update the landing page link in the Navbar component to point to the `/ento/opencv` route instead of `/ento/calculator`. This change aligns the link with the updated page structure and improves the user experience.

Remove unused code in Logo component

The code changes remove unused code in the Logo component, improving code cleanliness and maintainability.

Update welcome message in Dashboard component

The code changes update the welcome message in the Dashboard component to include the user's username. This change personalizes the greeting and enhances the user experience.

Update pageExtensions in next.config.js and fix broken links

The code changes update the `next.config.js` file to include the file extensions `ts` and `tsx` in the `pageExtensions` array. This change enables client-side rendering for the council detail components. Additionally, broken links in the codebase have been fixed, improving the user experience.

Note: This suggestion assumes that the recent repository commits follow the conventional format of using a prefix (e.g., "feat", "fix", "refactor") followed by a colon to indicate the type of change.

+++ app/shared/components/council-management/councilManagement.tsx
@@ -19,8 +19,6 @@ import { useEffect, useState } from "react";
 import useSWR from "swr";
 import ErrorComponent from "../error";
 import LoadingComponent from "../loading";
-import PageHeader from "../general-component/page-component/PageHeader";
-
 type CouncilListProps = {
 	id: number;
 	name: string;
@@ -30,6 +28,7 @@ type CouncilListProps = {
 	leaderEmail: string;
 	createdAt: Date;
 	createdBy: string;
+	leaderId: number;
 };
 
 export default function CouncilManagement() {
@@ -52,7 +51,10 @@ export default function CouncilManagement() {
 
 	useEffect(() => {
 		if (councilResponse && councilResponse.data) {
-			setCouncil(councilResponse.data);
+			setCouncil({
+				...councilResponse.data,
+				leaderEmail: councilResponse.data.leader.email,
+			});
 		}
 	}, [councilResponse]);
 
@@ -94,17 +96,13 @@ export default function CouncilManagement() {
 
 	const handleSave = async () => {
 		setIsLoading(true);
-		const updatedCouncil = {
-			...council,
-			councilId: userData.councilId,
-		};
 		const res = await fetch("/api/council/update", {
 			method: "PUT",
 			headers: {
 				"Content-Type": "application/json",
 			},
 			body: JSON.stringify({
-				council: updatedCouncil,
+				council: council,
 			}),
 		}).then((res) => res.json());
 
@@ -119,28 +117,32 @@ export default function CouncilManagement() {
 	};
 
 	const handleDelete = async () => {
-		setIsLoading(true);
-		const res = await fetch(
-			`/api/council/delete?councilId=${userData.councilId}`,
-			{
-				method: "DELETE",
-				headers: {
-					"Content-Type": "application/json",
-				},
-			}
-		).then((res) => res.json());
+		if (window.confirm("Are you sure you want to quit the council?")) {
+			setIsLoading(true);
+			const res = await fetch(
+				`/api/council/delete?councilId=${userData.councilId}`,
+				{
+					method: "DELETE",
+					headers: {
+						"Content-Type": "application/json",
+					},
+				}
+			).then((res) => res.json());
 
-		mutateCouncilRes();
-		mutateUser();
-		toast({
-			title: res.status === 200 ? res.message : res.error,
-			status: res.status === 200 ? "success" : "error",
-			duration: 3000,
-			isClosable: true,
-			position: "bottom-right",
-		});
+			toast({
+				title: res.status === 200 ? res.message : res.error,
+				status: res.status === 200 ? "success" : "error",
+				duration: 3000,
+				isClosable: true,
+				position: "bottom-right",
+			});
+			if (res.status === 200) {
+				mutateCouncilRes();
+				mutateUser();
+				router.push("/council/new");
+			}
+		}
 		setIsLoading(false);
-		router.push("/council/new");
 	};
 
 	return (
